# Conflicts:
#	Phase1/ProgressReport.md

+++ main/Entities/InterviewIntern.java
@@ -11,7 +11,7 @@ public class InterviewIntern extends Intern{
 
     private String interviewInternName;
     private int interviewInternAge;
-    private HashMap<String, Integer> interviewInternSkills;
+    private HashMap<String, Double> interviewInternSkills;
     private ResponseTree<ArrayList<String>> responseTree;
 
     /**
@@ -23,7 +23,7 @@ public class InterviewIntern extends Intern{
      * @param interviewInternSkills A map of skills where the key is the name of the skill and the value is the percentage.
      */
     public InterviewIntern(String interviewInternName, int interviewInternAge,
-                           HashMap<String, Integer> interviewInternSkills){
+                           HashMap<String, Double> interviewInternSkills){
         super(interviewInternName, interviewInternAge, interviewInternSkills);
     }
