+++ test/UseCases/ResponseTreeMakerTest.java
@@ -14,10 +14,10 @@ public class ResponseTreeMakerTest {
 
     @Test(timeout = 1000)
     public void TestGenerateInternResponses() throws FileNotFoundException {
-        HashMap<String, Integer> skills = new HashMap<>();
-        skills.put("Confidence", 50);
-        skills.put("Teamwork", 50);
-        skills.put("Flexibility", 50);
+        HashMap<String, Double> skills = new HashMap<>();
+        skills.put("Confidence", 50.0);
+        skills.put("Teamwork", 50.0);
+        skills.put("Flexibility", 50.0);
         InterviewIntern intern = new InterviewIntern("Camille", 20, skills);
         ResponseTreeMaker rtm = new ResponseTreeMaker(intern);
         ResponseTree<ArrayList<String>> respTree = rtm.generateInternResponses();
@@ -26,10 +26,10 @@ public class ResponseTreeMakerTest {
 
     @Test(timeout = 1000)
     public void TestAssignResponseToIntern() throws FileNotFoundException {
-        HashMap<String, Integer> skills = new HashMap<>();
-        skills.put("Confidence", 50);
-        skills.put("Teamwork", 50);
-        skills.put("Flexibility", 50);
+        HashMap<String, Double> skills = new HashMap<>();
+        skills.put("Confidence", 50.0);
+        skills.put("Teamwork", 50.0);
+        skills.put("Flexibility", 50.0);
         InterviewIntern intern = new InterviewIntern("Camille", 20, skills);
         ResponseTreeMaker rtm = new ResponseTreeMaker(intern);
         ResponseTree<ArrayList<String>> respTree = rtm.generateInternResponses();
